initial commit
Snake Water Gun is a game in which you have to choose any one element among this(s for Snake, w for Water and g for Gun). Here, Snake will win in front of water, Water will win in front of Gun and Gun will win in front of Snake.